requirements.txt
openai
RPi.GPIO
python-dotenv
requests
